Improve documentation of accepted schema
Although the intention of the current documentation seems to be to list
implementation details of this validator, it serves partially or can be
confused as what is currently supported at all.

Extend the documentation to describe all constraints supported, so that
this is more clear. More so given that the upstream schema definition
is now gone, and is only available via the archive.org.